+++ feature/eventeditor/api/src/main/java/com/sergei/pokhodai/expensemanagement/feature/eventeditor/api/router/EventEditorRouter.kt
@@ -1,5 +1,5 @@
 package com.sergei.pokhodai.expensemanagement.feature.eventeditor.api.router
 
 interface EventEditorRouter {
-    fun goToEventEditor(eventId: Int?)
+    fun goToEventEditor(eventId: Long?)
 }
