+++ feature/eventeditor/impl/src/main/java/com/sergei/pokhodai/expensemanagement/feature/eventeditor/impl/router/contract/EventEditorRouterContract.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 internal data class EventEditorRouterContract(
-    val eventId: Int?
+    val eventId: Long?
 ) : RouterContract
