+++ feature/user/impl/src/main/java/com/sergei/pokhodai/expensemanagement/feature/user/impl/router/contract/UserRouterContract.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 internal data class UserRouterContract(
-    val userId: Int? = null
+    val userId: Long? = null
 ) : RouterContract
